+++ cpp/src/binaryop/jit/operation-udf.hpp
@@ -1,8 +1,5 @@
 /*
- * Copyright (c) 2019-2020, NVIDIA CORPORATION.
- *
- * Copyright 2018-2019 BlazingDB, Inc.
- *     Copyright 2018 Christian Noboa Mardini <christian@blazingdb.com>
+ * Copyright (c) 2021, NVIDIA CORPORATION.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -19,21 +16,5 @@
 
 #pragma once
 
-namespace cudf {
-namespace rolling {
-namespace jit {
-namespace code {
-extern const char* kernel_headers;
-extern const char* kernel;
-extern const char* operation_h;
-
-extern const char* kernel_headers;
-extern const char* kernel;
-extern const char* operation_h;
-
-extern const char* grouped_window_wrapper;
-
-}  // namespace code
-}  // namespace jit
-}  // namespace rolling
-}  // namespace cudf
+// This file serves as a placeholder for user defined functions, so jitify can choose to override it
+// at runtime.
