+++ cpp/src/copying/pack.cpp
@@ -17,8 +17,6 @@
 #include <cudf/detail/copy.hpp>
 #include <cudf/detail/nvtx/ranges.hpp>
 
-#include <jit/type.h>
-
 namespace cudf {
 namespace detail {
