+++ cpp/src/transform/transform.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019, NVIDIA CORPORATION.
+ * Copyright (c) 2019-2021, NVIDIA CORPORATION.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,12 +14,11 @@
  * limitations under the License.
  */
 
-#include "jit/code/code.h"
+#include <jit_preprocessed_files/transform/jit/kernel.cu.jit.hpp>
 
-#include <jit/launcher.h>
-#include <jit/parser.h>
-#include <jit/type.h>
-#include <jit/common_headers.hpp>
+#include <jit/cache.hpp>
+#include <jit/parser.hpp>
+#include <jit/type.hpp>
 
 #include <cudf/column/column.hpp>
 #include <cudf/column/column_factories.hpp>
@@ -29,26 +28,11 @@
 #include <cudf/utilities/traits.hpp>
 #include <cudf/utilities/type_dispatcher.hpp>
 
-#include <jit/timestamps.hpp.jit>
-#include <jit/types.hpp.jit>
-
 #include <rmm/cuda_stream_view.hpp>
 
 namespace cudf {
 namespace transformation {
-//! Jit functions
 namespace jit {
-
-const std::vector<std::string> header_names{cudf_types_hpp, cudf_wrappers_timestamps_hpp};
-
-std::istream* headers_code(std::string filename, std::iostream& stream)
-{
-  auto it = cudf::jit::stringified_headers.find(filename);
-  if (it != cudf::jit::stringified_headers.end()) {
-    return cudf::jit::send_stringified_header(stream, it->second);
-  }
-  return nullptr;
-}
 
 void unary_operation(mutable_column_view output,
                      column_view input,
@@ -57,28 +41,25 @@ void unary_operation(mutable_column_view output,
                      bool is_ptx,
                      rmm::cuda_stream_view stream)
 {
-  std::string hash = "prog_transform" + std::to_string(std::hash<std::string>{}(udf));
+  std::string kernel_name =
+    jitify2::reflection::Template("cudf::transformation::jit::kernel")  //
+      .instantiate(cudf::jit::get_type_name(output.type()),  // list of template arguments
+                   cudf::jit::get_type_name(input.type()));
 
-  std::string cuda_source = code::kernel_header;
-  if (is_ptx) {
-    cuda_source += cudf::jit::parse_single_function_ptx(
-                     udf, "GENERIC_UNARY_OP", cudf::jit::get_type_name(output_type), {0}) +
-                   code::kernel;
-  } else {
-    cuda_source += cudf::jit::parse_single_function_cuda(udf, "GENERIC_UNARY_OP") + code::kernel;
-  }
+  std::string cuda_source =
+    is_ptx ? cudf::jit::parse_single_function_ptx(udf,  //
+                                                  "GENERIC_UNARY_OP",
+                                                  cudf::jit::get_type_name(output_type),
+                                                  {0})
+           : cudf::jit::parse_single_function_cuda(udf,  //
+                                                   "GENERIC_UNARY_OP");
 
-  // Launch the jitify kernel
-  cudf::jit::launcher(hash,
-                      cuda_source,
-                      header_names,
-                      cudf::jit::compiler_flags,
-                      headers_code,
-                      stream)
-    .set_kernel_inst("kernel",  // name of the kernel we are launching
-                     {cudf::jit::get_type_name(output.type()),  // list of template arguments
-                      cudf::jit::get_type_name(input.type())})
-    .launch(output.size(), cudf::jit::get_data_ptr(output), cudf::jit::get_data_ptr(input));
+  cudf::jit::get_program_cache(*transform_jit_kernel_cu_jit)
+    .get_kernel(kernel_name, {}, {{"transform/jit/operation-udf.hpp", cuda_source}})  //
+    ->configure_1d_max_occupancy(0, 0, 0, stream.value())                             //
+    ->launch(output.size(),                                                           //
+             cudf::jit::get_data_ptr(output),
+             cudf::jit::get_data_ptr(input));
 }
 
 }  // namespace jit
